+++ broker/src/main/resources/db/migration/V1_0_1__quartz_tables_mysql_innodb.sql
@@ -1,5 +1,4 @@
-#
-# In your Quartz properties file, you'll need to set 
+# In your Quartz properties file, you'll need to set
 # org.quartz.jobStore.driverDelegateClass = org.quartz.impl.jdbcjobstore.StdJDBCDelegate
 #
 #
@@ -152,6 +151,8 @@ CREATE TABLE IF NOT EXISTS QRTZ_LOCKS (
   PRIMARY KEY (SCHED_NAME, LOCK_NAME)
 )
   ENGINE = InnoDB;
+
+# Modified by Swisscom (Schweiz) AG) on 23th of May 2017
 
 CREATE INDEX IDX_QRTZ_J_REQ_RECOVERY
   ON QRTZ_JOB_DETAILS (SCHED_NAME, REQUESTS_RECOVERY);